Refactor Parser.parseAssignExpression.
